Week 4: User entity + Controller + Service, One to Many relationship, Cascade types, and query to fetch books by user

+++ src/main/java/com/example/demo/controller/BookController.java
@@ -34,19 +34,19 @@ public class BookController {
 
     //READ ONE
     @GetMapping("/{id}")
-    public Book getBookById(@PathVariable int id) {
+    public Book getBookById(@PathVariable Long id) {
         return bookService.getBookById(id);
     }
 
     //UPDATE
     @PutMapping("/{id}")
-    public Book updateBookById(@PathVariable int id, @RequestBody Book bookDetails) {
+    public Book updateBookById(@PathVariable Long id, @RequestBody Book bookDetails) {
         return bookService.updateBookById(id, bookDetails);
     }
 
     //DELETE
     @DeleteMapping("/{id}")
-    public String deleteBook(@PathVariable int id) {
+    public String deleteBook(@PathVariable Long id) {
         return bookService.deleteBookById(id);
     }
